pyproject.toml
[tool.poetry]
name = "ord-plan"
version = "0.0.1"
description = "ORD Plan"
authors = ["vonpupp <vonpupp@gmail.com>"]
license = "GPL-3.0"
readme = "README.md"
homepage = "https://github.com/vonpupp/ord-plan"
repository = "https://github.com/vonpupp/ord-plan"
documentation = "https://ord-plan.readthedocs.io"
classifiers = [
    "Development Status :: 3 - Alpha",
]

[tool.poetry.urls]
Changelog = "https://github.com/vonpupp/releases"

[tool.poetry.dependencies]
python = "^3.8"
click = ">=8.0.1"
orgparse = ">=0.1.0"
python-dateutil = ">=2.8.0"
croniter = ">=0.3.0"
PyYAML = ">=6.0"

[tool.poetry.group.dev.dependencies]
Pygments = ">=2.10.0"
black = ">=21.10b0"
coverage = {extras = ["toml"], version = ">=6.2,<7.0"}
darglint = ">=1.8.1"
flake8 = ">=4.0.1"
flake8-bandit = ">=2.1.2"
flake8-bugbear = ">=21.9.2"
flake8-docstrings = ">=1.6.0"
flake8-rst-docstrings = ">=0.2.5"
furo = ">=2021.11.12"
isort = ">=5.10.1"
mypy = ">=0.930"
pep8-naming = ">=0.12.1"
pre-commit = ">=2.16.0"
pre-commit-hooks = ">=4.1.0"
pytest = ">=6.2.5"
pytest-cov = ">=4.0.0"
types-python-dateutil = ">=2.8.0"
types-PyYAML = ">=6.0.0"
types-croniter = ">=2.0.0"

pyupgrade = ">=2.29.1"
safety = ">=1.10.3"
sphinx = ">=5.0.0"
sphinx-autobuild = ">=2021.3.14"
sphinx-click = ">=3.0.2"
# typeguard = ">=2.13.3"  # Temporarily disabled due to CLI issues
xdoctest = {extras = ["colors"], version = ">=0.15.10"}
myst-parser = {version = ">=0.16.1"}

[tool.poetry.scripts]
ord-plan = "ord_plan.__main__:main"

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["ord_plan", "tests"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning:pytz.*:",
    "ignore::DeprecationWarning:_pytest.*:",
]

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2

[tool.mypy]
strict = true
warn_unreachable = false
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
